+++ tests/functional-tests/tests/consistency_group_test.py
@@ -1,6 +1,6 @@
 import logging
 import pytest
-import ibm_spectrum_scale_csi.scale_operator as scaleop
+import ibm_spectrum_scale_csi.base_class as baseclass
 import ibm_spectrum_scale_csi.common_utils.input_data_functions as inputfunc
 LOGGER = logging.getLogger()
 pytestmark = [pytest.mark.volumeprovisioning, pytest.mark.localcluster, pytest.mark.cg]
@@ -31,7 +31,7 @@ def values(request, check_csi_operator):
         value_pvc = [{"access_modes": "ReadWriteMany", "storage": "1Gi"}]
         value_pod = [{"mount_path": "/usr/share/nginx/html/scale", "read_only": "False"}]
 
-    driver_object = scaleop.Driver(kubeconfig_value, value_pvc, value_pod, data["id"],
+    driver_object = baseclass.Driver(kubeconfig_value, value_pvc, value_pod, data["id"],
                                    test_namespace, keep_objects, data["image_name"], data["pluginNodeSelector"])
 
 
@@ -40,10 +40,10 @@ def values(request, check_csi_operator):
 def test_get_version():
     LOGGER.info("Cluster Details:")
     LOGGER.info("----------------")
-    scaleop.filesetfunc.get_scale_version(data)
-    scaleop.kubeobjectfunc.get_kubernetes_version(kubeconfig_value)
-    scaleop.kubeobjectfunc.get_operator_image()
-    scaleop.kubeobjectfunc.get_driver_image()
+    baseclass.filesetfunc.get_scale_version(data)
+    baseclass.kubeobjectfunc.get_kubernetes_version(kubeconfig_value)
+    baseclass.kubeobjectfunc.get_operator_image()
+    baseclass.kubeobjectfunc.get_driver_image()
 
 
 def test_driver_cg_pass_1():
